requirements.txt
pandas>=1.3.0
thefuzz>=0.19.0
flashtext>=2.7
camelot-py[cv]>=0.11.0
python-dateutil>=2.8.0
datefinder>=0.7.0
pyinaturalist>=0.18.0
requests>=2.26.0